First World Progress
-Updated base scene and added the first scene inside the castle.
-Will add scenery and make it look better later. This is focused on a playable demo.

SCENES:

BASE: The main camera has a clamped background that follows the player. If the player falls below a certain y-axis, he dies. No threats in base level, just walk up to the castle.

LEVEL1: Start inside by the door that you just entered through. Jump up the stairs and the player encounters a locked door. Continuing along, the player meets the first threat- a pit with spikes at the bottom. If you fail the jump, you die. The player meets two sets of spikes after that as well. Continue along the path and the player finds the first enemy. If the player jumps on the enemy, the enemy dies. If the enemy hits the player, the player dies. Push the crate into position to jump and reach the level that opens the locked door. Return to the door to open it and advance to the next scene.

TODO (Base): Expand the level, add some scenery and the player's house in background. Maybe add a require jump obstacle. Need to make a way to open the doors that triggers a load of the next scene.

TODO(Level1): Add colliders to the ground spikes and make a trigger event that kills the player if they don't successfully jump over them. Add a script that allows the player to use the level to unlock the door.

TODO: Add next scene, potentially final level 1 scene that is longer and more difficult. Brainstorm more puzzles or mechanics that we can integrate. Add scripts to showcase the players unique powers.